services/connectors/readwise/pyproject.toml
[project]
name = "connector-readwise"
version = "0.1.0"
description = "Readwise connector"
requires-python = ">=3.12"
dependencies = [
  "httpx>=0.27.0",
  "python-dateutil>=2.9.0",
  "readwise-api @ git+https://github.com/reuteras/readwise-api",
]

[tool.uv]
package = true

[tool.hatch.build.targets.wheel]
packages = ["src/connector_readwise"]

[tool.hatch.metadata]
allow-direct-references = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
